backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-socketio==5.11.2
python-engineio==4.9.1
pandas==2.2.2
openpyxl==3.1.2
watchdog==4.0.0